requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
bertopic==0.17.0
blinker==1.9.0
blis==1.3.0
Brotli==1.1.0
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
cssselect==1.3.0
cssselect2==0.8.0
cycler==0.12.1
cymem==2.0.11
DAWG2-Python==0.9.0
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.12
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
Flask==3.1.0
fonttools==4.58.0
fsspec==2025.3.2
googlesearch-python==1.3.0
greenlet==3.2.2
h11==0.16.0
hdbscan==0.8.40
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.1
idna==3.10
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.44.0
lxml==5.4.0
lxml_html_clean==0.4.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
narwhals==1.39.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numba==0.61.2
numpy==2.2.5
openai==1.78.1
packaging==25.0
pandas==2.2.3
pdfkit==1.0.0
pillow==11.2.1
plotly==6.0.1
preshed==3.0.9
protobuf==6.31.0
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.1
pymorphy3==2.0.3
pymorphy3-dicts-ru==2.4.417150.4580142
pymorphy3-dicts-uk==2.4.1.1.1663094765
pynndescent==0.5.13
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==4.1.0
sentencepiece==0.2.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
starlette==0.46.2
sympy==1.14.0
Telethon==1.40.0
thinc==8.3.6
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uk_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/uk_core_news_sm-3.8.0/uk_core_news_sm-3.8.0-py3-none-any.whl
umap-learn==0.5.7
urllib3==2.4.0
uvicorn==0.34.2
wasabi==1.1.3
weasel==0.4.1
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
wordcloud==1.9.4
wrapt==1.17.2
zopfli==0.2.3.post1
